rename mods in tree
